refactors List to own file
